+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,8 +7,6 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
 
 class ReservationTest {
