+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservationTime;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,7 +6,6 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.domain.ReservationTime;
 
 public class ReservationTimeTest {
